move to proper (cut-down) implementation of get_order, and fix the argument to be unsigned long again

+++ defs.h
@@ -267,8 +267,8 @@ function (boolean) hyp_pool_wf (pointer pool_pointer, struct hyp_pool pool,
         || (range_end_virt <= ((u64) pool_pointer)))
 }
 
-function (u8) get_order_uf (u32 size) {
-  (u8) bw_fls_uf(shift_right(size - 1u32, 12u32))
+function (u8) get_order_uf (u64 size) {
+  (u8) bw_fls_uf(shift_right(size - 1u64, 12u64))
 }
 
 function (pointer) virt (pointer phys, i64 physvirt_offset) {
